add several changes – rename DataProvider to SPODataProvider – move interface definition (model) in separate file – adjust references accordingly
